Move EndTimestamp into it's own file
